clustermesh: move MCS-API config to types package
Move MCSAPIConfig to types package to prevent future dependency cycle.
Also introduce DefaultMCSAPIConfig symbol at the same time.
