+++ pkg/clustermesh/mcsapi/cell.go
@@ -17,6 +17,7 @@ import (
 	ctrlRuntime "sigs.k8s.io/controller-runtime"
 	mcsapiv1alpha1 "sigs.k8s.io/mcs-api/pkg/apis/v1alpha1"
 
+	mcsapitypes "github.com/cilium/cilium/pkg/clustermesh/mcsapi/types"
 	"github.com/cilium/cilium/pkg/clustermesh/operator"
 	"github.com/cilium/cilium/pkg/clustermesh/types"
 	k8sClient "github.com/cilium/cilium/pkg/k8s/client"
@@ -44,7 +45,7 @@ type mcsAPIParams struct {
 	AgentConfig *option.DaemonConfig
 	ClusterMesh operator.ClusterMesh
 	Cfg         operator.ClusterMeshConfig
-	CfgMCSAPI   operator.MCSAPIConfig
+	CfgMCSAPI   mcsapitypes.MCSAPIConfig
 
 	// ClusterInfo is the id/name of the local cluster.
 	ClusterInfo types.ClusterInfo
