+++ pkg/clustermesh/mcsapi/serviceexportsync_test.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/cilium/cilium/daemon/k8s"
 	"github.com/cilium/cilium/pkg/clustermesh/mcsapi/types"
-	"github.com/cilium/cilium/pkg/clustermesh/operator"
+	mcsapitypes "github.com/cilium/cilium/pkg/clustermesh/mcsapi/types"
 	envoyCfg "github.com/cilium/cilium/pkg/envoy/config"
 	"github.com/cilium/cilium/pkg/hive"
 	k8sFakeClient "github.com/cilium/cilium/pkg/k8s/client/testutils"
@@ -115,8 +115,8 @@ func Test_mcsServiceExportSync_Reconcile(t *testing.T) {
 		k8s.ResourcesCell,
 		cell.Config(envoyCfg.SecretSyncConfig{}),
 		cell.Provide(ServiceExportResource),
-		cell.Provide(func() operator.MCSAPIConfig {
-			return operator.MCSAPIConfig{ClusterMeshEnableMCSAPI: true}
+		cell.Provide(func() mcsapitypes.MCSAPIConfig {
+			return mcsapitypes.MCSAPIConfig{ClusterMeshEnableMCSAPI: true}
 		}),
 		cell.Invoke(func(
 			svc resource.Resource[*slim_corev1.Service],
