+++ pkg/clustermesh/operator/clustermesh.go
@@ -30,7 +30,7 @@ type clusterMesh struct {
 	common common.ClusterMesh
 
 	cfg       ClusterMeshConfig
-	cfgMCSAPI MCSAPIConfig
+	cfgMCSAPI mcsapitypes.MCSAPIConfig
 	logger    *slog.Logger
 	metrics   Metrics
